users/serializers.py
from users.models import Users
from rest_framework import serializers


class UsersSerializer(serializers.ModelSerializer):
    class Meta:
        model = Users
        fields = '__all__'

    def create(self, validated_data):
        users = Users.objects.create_user(**validated_data)
        return users